requirements.txt
flask
flask-migrate
flask-smorest
gunicorn
python-dotenv
flask-sqlalchemy
flask-jwt-extended
passlib
alembic
psycopg2
requests
rq
redis